% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_format_data.R
\name{aggregate_format_data}
\alias{aggregate_format_data}
\title{Aggregate and format transect summary data}
\usage{
aggregate_format_data <- function(projecthome) {}
}
\arguments{
\item{projecthome}{Project directory normally obtained using `getwd()`}
}
\value{
The function will export three .csv files with data summarized by
BASIN, STRATUM, GRID, LAYER (epilimnion vs. hypolimnion region), and interval format.
'hacdat.csv' includes  integration, mean target strength, and fish density estimates.
'histo.csv' includes single target counts for 0.5 dB target strength bins ranging
from -64.5 to -20.5 dB. 'histohac.csv' merges 'hacdat' and 'histo' together.
}
\description{
Provides code to import, aggregate, and format integration
and target strength data from hydroacoustic survey.
}
\examples{
\dontrun{
}
}
