% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_Import_EVcsv.R
\name{Analysis_Import_EVcsv}
\alias{Analysis_Import_EVcsv}
\title{Import Echoview Analysis Exports}
\usage{
Analysis_Import_EVcsv()
}
\value{
a list with intg, ts and histo data
}
\description{
Import Echoview Analysis Exports
}
\examples{
\dontrun{
library(erieacoustics)
hacdat <- Analysis_Import_EVcsv()
 }
}
