% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Effort_Allocation_Data.R
\name{Effort_Allocation}
\alias{Effort_Allocation}
\title{Effort Allocation}
\value{
dataframe
}
\description{
This table contains the effort allocation data for each strata.
}
\examples{
data("Effort_Allocation")
head(Effort_Allocation)
}
