% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_transect_evdata.R
\name{export_transect}
\alias{export_transect}
\title{Export Echoview transect data}
\usage{
export_transect(prjdir, transectname, horizbin)
}
\arguments{
\item{horizbin}{}
}
\value{
The function will attempt to return a cruise track (`gps.csv`), a copy of the
edited epi line in the Echoview format (`EpiLayerLine_Edited.evl`) and as a csv
(`EpiLayerLine_Final.csv`)
}
\description{
Provides a series of COM commands to set export parameters and then export
the required analysis files.
}
\examples{
\dontrun{
library(erieacoustics)
# export_transect(getwd(), "R21_S22", 800)
}
}
