% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finish_setup().R
\name{finish_setup}
\alias{finish_setup}
\title{Finish erieacoustic project setup}
\usage{
finish_setup()
}
\value{

}
\description{
Several of the analysis procedures rely on custom functions. Rather than a
vignette approach, this function copies template analysis scripts that establish a workflow
based on the existing package functions. The template files are "auto-magically" created
following the project naming procedures and directories.
}
\examples{
finish_setup()
}
