% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_epi_bottom_lines.R
\name{import_epi_bottom_lines}
\alias{import_epi_bottom_lines}
\title{import_epi_bottom_lines}
\usage{
import_epi_bottom_lines(transect)
}
\arguments{
\item{transect}{A transect in the expected `erieacoustics` format ("EB_S16_G1004")}
}
\value{
summary of epi and bottom line data from Echoview
}
\description{
A helper function to import, format and summarize the epi and bottom lines
that are exported from Echoview
}
\examples{
\dontrun{
library(erieacoustics)
library(dplyr)
import_epi_bottom_lines("EB_S16_G1004")

alltransects <- dir("3_Ping_Data")
epi_bot_lines <- bind_rows(lapply(alltransects, import_epi_bottom_lines))
}
}
