% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grids_proposed.R
\name{sample_grid_rand_select}
\alias{sample_grid_rand_select}
\title{Randomly generate proposed sample grids}
\usage{
sample_grid_rand_select(basin, year)
}
\arguments{
\item{basin}{specify the basin (ex. EB, CB or WB)}

\item{year}{intended survey year}
}
\value{
sample_grids_proposed.png, sample_grids_proposed.csv,
and sample_grids_final.csv
}
\description{
This function randomly selects sample grids for each strata
based on effort allocation table.
}
\examples{
\dontrun{
sample_grid_rand_select("WB",2022)
 }
}
