% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grids_final.R
\name{sample_grids_final}
\alias{sample_grids_final}
\title{Finalize sample grids}
\usage{
sample_grids_final(basin, year)
}
\arguments{
\item{basin}{specify the basin (ex. EB, CB or WB)}

\item{year}{intended survey year}
}
\value{
sample_grids_final.png
}
\description{
This function creates final survey may pulling
from sample_grids_final.csv.
}
\examples{
\dontrun{
sample_grids_final("WB",2022)
 }
}
