% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_transect.R
\name{set_up_transect}
\alias{set_up_transect}
\title{Automate the setup of a template}
\usage{
set_up_transect(template, projecthome, sonartype, tranname)
}
\arguments{
\item{template}{A file path to an Echoview template}

\item{projecthome}{A file path to the project directory. Generally, when working in an
RStudio project environment you can populate this field using `getwd()`}

\item{sonartype}{The function uses a text pattern matching to find acoustic data files.
Function expects an input of *BIOSONIC* or *SIMRAD* to detect \*.dt4 or \*.raw
file types.}

\item{tranname}{Transect name that matches a valid directory in *PINGS/* directory}
}
\value{

}
\description{
This function calls RDCOMClient to establish a COM application for Echoview.
This function is primarily a wrapper for functions found in EchoviewR
(https://github.com/AustralianAntarcticDivision/EchoviewR) applied to files stored
in an expected file path.
}
\examples{
\dontrun{
library(erieacoustics)
evtemplate <- file.path(getwd(), 'EVTemplate/EVTemplate.EV')
dir.exists("Pings")
file.exists(evtemplate)

set_up_transect(evtemplate, projecthome = getwd(),
               sonartype = "SIMRAD", tranname = "ERIE")
set_up_transect(evtemplate, projecthome = getwd(),
             sonartype = "BIOSONIC", tranname = "COB")
}

}
