% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trawl_Depths_Data.R
\name{Trawl_Depths}
\alias{Trawl_Depths}
\title{Trawl Depths}
\value{
dataframe
}
\description{
This table contains the target trawl depths for each site by strata.
}
\examples{
data("Trawl_Depths")
head(Trawl_Depths)
}
